add comments to new PR, set assignee
