+++ on-pull-request-runner/src/lib/git.ts
@@ -53,6 +53,11 @@ export function configureGitIdentity(cwd: string): void {
   git(['config', 'user.email', email], cwd);
 }
 
+/** Resolves a git ref to a full commit SHA. */
+export function gitRevParse(ref: string, cwd: string): string {
+  return git(['rev-parse', ref], cwd);
+}
+
 /** Creates or resets a branch at the current HEAD. */
 export function gitCheckoutBranch(branch: string, cwd: string): void {
   git(['checkout', '-B', branch], cwd);
