+++ on-pull-request-runner/src/open-remediation-pr.ts
@@ -3,14 +3,18 @@
  */
 import fs from 'node:fs';
 import yaml from 'yaml';
+import { publishAuditFeedback } from './lib/audit-feedback.js';
 import { applyOrlFixes } from './lib/apply-orl-fixes.js';
 import { artifactPath } from './lib/artifacts.js';
+import { envInt } from './lib/env.js';
 import {
   configureGitIdentity,
   gitAddAll,
   gitCheckoutBranch,
   gitCommit,
+  gitDiffNameOnly,
   gitPush,
+  gitRevParse,
   gitStatusPorcelain,
 } from './lib/git.js';
 import { GitHubClient, parseOwnerRepo } from './lib/github-client.js';
@@ -20,10 +24,14 @@ import { totalsFromBatchReports } from './lib/report-counts.js';
 import { runMain } from './lib/runner.js';
 import type { EvaluationBatch, OrlReport } from './types.js';
 
+function loadJson<T>(file: string): T {
+  return JSON.parse(fs.readFileSync(file, 'utf8')) as T;
+}
+
 function loadBatches(): EvaluationBatch[] {
-  const raw = JSON.parse(
-    fs.readFileSync(artifactPath('evaluation-batches.json'), 'utf8')
-  ) as { batches: EvaluationBatch[] };
+  const raw = loadJson<{ batches: EvaluationBatch[] }>(
+    artifactPath('evaluation-batches.json')
+  );
   return raw.batches;
 }
 
@@ -50,9 +58,7 @@ function loadBatchReport(batchId: string): OrlReport | null {
 function loadStagedFiles(batchId: string): string[] | null {
   const manifestPath = artifactPath(`batches/${batchId}/staged-files.json`);
   if (!fs.existsSync(manifestPath)) return null;
-  const raw = JSON.parse(fs.readFileSync(manifestPath, 'utf8')) as {
-    files?: string[];
-  };
+  const raw = loadJson<{ files?: string[] }>(manifestPath);
   return raw.files ?? null;
 }
 
@@ -77,6 +83,57 @@ async function findOpenRemediationPr(args: {
     number: match.number,
     html_url: `https://github.com/${args.owner}/${args.repo}/pull/${match.number}`,
   };
+}
+
+function remediationScannableFiles(args: {
+  copiedPaths: string[];
+  workspaceRoot: string;
+  baseSha: string;
+  headSha: string;
+}): string[] {
+  const fromDiff = gitDiffNameOnly({
+    baseSha: args.baseSha,
+    headSha: args.headSha,
+    cwd: args.workspaceRoot,
+  });
+  return [...new Set([...args.copiedPaths, ...fromDiff])].sort();
+}
+
+async function publishRemediationFeedback(args: {
+  github: GitHubClient;
+  owner: string;
+  repo: string;
+  remediationPullNumber: number;
+  workspaceRoot: string;
+  baseSha: string;
+  headSha: string;
+  scannableFiles: string[];
+  sourcePullNumber: number;
+  sourceHeadRef: string;
+}): Promise<void> {
+  const portalServiceUrl = (
+    process.env.INPUT_PORTAL_SERVICE_URL?.trim() || 'https://app.gomboc.ai'
+  ).replace(/\/+$/, '');
+  const maxComments = envInt('INPUT_COMMENT_MAX_PER_PR', 50);
+
+  await publishAuditFeedback({
+    github: args.github,
+    owner: args.owner,
+    repo: args.repo,
+    pullNumber: args.remediationPullNumber,
+    headSha: args.headSha,
+    baseSha: args.baseSha,
+    workspaceRoot: args.workspaceRoot,
+    scannableFiles: args.scannableFiles,
+    portalServiceUrl,
+    maxComments,
+    summaryTarget: 'pull_body',
+    introLines: [
+      `Automated ORL remediation stacked on \`${args.sourceHeadRef}\` for PR #${args.sourcePullNumber}.`,
+      '',
+      'Review and merge this PR into your feature branch before merging the original PR.',
+    ],
+  });
 }
 
 async function main(): Promise<void> {
@@ -153,6 +210,15 @@ async function main(): Promise<void> {
   gitPush('origin', botBranch, workspaceRoot);
   console.log(`Pushed remediation branch ${botBranch}`);
 
+  const headSha = gitRevParse('HEAD', workspaceRoot);
+  const baseSha = gitRevParse('HEAD~1', workspaceRoot);
+  const scannableFiles = remediationScannableFiles({
+    copiedPaths,
+    workspaceRoot,
+    baseSha,
+    headSha,
+  });
+
   const { owner, repo } = parseOwnerRepo(pr.repository);
   const github = GitHubClient.fromEnv();
 
@@ -163,27 +229,62 @@ async function main(): Promise<void> {
     headRef: botBranch,
   });
 
-  if (existing) {
-    console.log(
-      `Updated existing remediation PR #${existing.number}: ${existing.html_url}`
-    );
-    return;
-  }
+  let remediationPullNumber: number;
+  let remediationUrl: string;
 
-  const created = await github.createPullRequest({
-    owner,
-    repo,
-    title: commitMessage,
-    head: botBranch,
-    base: pr.headRef,
-    body: [
+  if (existing) {
+    remediationPullNumber = existing.number;
+    remediationUrl = existing.html_url;
+    console.log(`Updated existing remediation PR #${remediationPullNumber}: ${remediationUrl}`);
+  } else {
+    const placeholderBody = [
       `Automated ORL remediation stacked on \`${pr.headRef}\` for PR #${pr.number}.`,
       '',
-      'Review and merge this PR into your feature branch before merging the original PR.',
-    ].join('\n'),
-  });
+      '_Assessment summary loading…_',
+    ].join('\n');
 
-  console.log(`Opened remediation PR #${created.number}: ${created.html_url}`);
+    const created = await github.createPullRequest({
+      owner,
+      repo,
+      title: commitMessage,
+      head: botBranch,
+      base: pr.headRef,
+      body: placeholderBody,
+    });
+    remediationPullNumber = created.number;
+    remediationUrl = created.html_url;
+    console.log(`Opened remediation PR #${remediationPullNumber}: ${remediationUrl}`);
+  }
+
+  if (pr.authorLogin) {
+    try {
+      await github.assignIssueAssignees({
+        owner,
+        repo,
+        issueNumber: remediationPullNumber,
+        assignees: [pr.authorLogin],
+      });
+      console.log(`Assigned remediation PR #${remediationPullNumber} to @${pr.authorLogin}`);
+    } catch (err) {
+      const message = err instanceof Error ? err.message : String(err);
+      console.warn(
+        `Could not assign remediation PR #${remediationPullNumber} to @${pr.authorLogin}: ${message}`
+      );
+    }
+  }
+
+  await publishRemediationFeedback({
+    github,
+    owner,
+    repo,
+    remediationPullNumber,
+    workspaceRoot,
+    baseSha,
+    headSha,
+    scannableFiles,
+    sourcePullNumber: pr.number,
+    sourceHeadRef: pr.headRef,
+  });
 }
 
 runMain(main);
